# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/files/{filename}": {
            "get": {
                "description": "Baixa um arquivo pelo nome",
                "produces": [
                    "application/octet-stream"
                ],
                "summary": "Faz o download de um arquivo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome do arquivo",
                        "name": "filename",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Arquivo baixado com sucesso",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "404": {
                        "description": "Arquivo não encontrado",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove permanentemente um arquivo armazenado",
                "summary": "Exclui um arquivo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome do arquivo",
                        "name": "filename",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Arquivo excluído com sucesso",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Arquivo não encontrado",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/upload": {
            "post": {
                "description": "Envia um arquivo e salva no volume do Docker",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "text/plain"
                ],
                "summary": "Faz upload de um arquivo",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Arquivo a ser enviado",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Arquivo salvo com sucesso",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Erro ao receber o arquivo",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro ao salvar arquivo",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}